catalog: move ConstraintDetail from descpb to catalog
This commit is a refactor in which we move descpb.ConstraintDetail to
the catalog package and replace some of its field types with catalog
interface types. This in turn makes it possible to break down the
DropConstraint method on tabledesc.Mutable so that it does not require
a callback for foreign key deletion.

Release note: None